src/test/java/seedu/address/ui/StatusBarFooterTest.java
package seedu.address.ui;

import static org.junit.Assert.assertEquals;
import static seedu.address.testutil.TypicalRequests.ALICE_REQUEST;
import static seedu.address.testutil.TypicalHealthWorkers.CARLIE;
import static seedu.address.ui.StatusBarFooter.SYNC_STATUS_INITIAL;
import static seedu.address.ui.StatusBarFooter.SYNC_STATUS_UPDATED;

import java.nio.file.Path;
import java.nio.file.Paths;
import java.text.SimpleDateFormat;
import java.time.Clock;
import java.time.Instant;
import java.time.ZoneId;
import java.util.Date;

import org.junit.AfterClass;
import org.junit.Before;
import org.junit.BeforeClass;
import org.junit.Test;

import guitests.guihandles.StatusBarFooterHandle;
import seedu.address.model.HealthWorkerBook;
import seedu.address.model.RequestBook;

public class StatusBarFooterTest extends GuiUnitTest {

    private static final Path STUB_SAVE_LOCATION = Paths.get("Stub");
    private static final Path RELATIVE_PATH = Paths.get(".");

    private static final Clock originalClock = StatusBarFooter.getClock();
    private static final Clock injectedClock = Clock.fixed(Instant.now(), ZoneId.systemDefault());

    private StatusBarFooterHandle statusBarFooterHandle;
    private final RequestBook requestBook = new RequestBook();
    private final HealthWorkerBook healthWorkerBook = new HealthWorkerBook();

    @BeforeClass
    public static void setUpBeforeClass() {
        // inject fixed clock
        StatusBarFooter.setClock(injectedClock);
    }

    @AfterClass
    public static void tearDownAfterClass() {
        // restore original clock
        StatusBarFooter.setClock(originalClock);
    }

    @Before
    public void setUp() {
        StatusBarFooter statusBarFooter = new StatusBarFooter(STUB_SAVE_LOCATION,
                STUB_SAVE_LOCATION, requestBook, healthWorkerBook);
        uiPartRule.setUiPart(statusBarFooter);

        statusBarFooterHandle = new StatusBarFooterHandle(statusBarFooter.getRoot());
    }

    @Test
    public void display() {
        // initial state, with only the initial sync status
        assertStatusBarContent("", SYNC_STATUS_INITIAL);

        // after request book is updated
        SimpleDateFormat formatter = new SimpleDateFormat("dd MMMM YYYY, hh:mm a");
        guiRobot.interact(() -> requestBook.addRequest(ALICE_REQUEST));
        long now = StatusBarFooter.getClock().millis();
        String lastUpdated = formatter.format(new Date(now));
        assertStatusBarContent(RELATIVE_PATH.resolve(STUB_SAVE_LOCATION).toString(), String.format(SYNC_STATUS_UPDATED,
                lastUpdated));

        // after health worker book is updated
        guiRobot.interact(() -> healthWorkerBook.addHealthWorker(CARLIE));
        now = StatusBarFooter.getClock().millis();
        lastUpdated = formatter.format(new Date(now));
        assertStatusBarContent(RELATIVE_PATH.resolve(STUB_SAVE_LOCATION).toString(), String.format(SYNC_STATUS_UPDATED,
                lastUpdated));
    }

    /**
     * Asserts that the save location matches that of {@code expectedSaveLocation}, and the
     * sync status matches that of {@code expectedSyncStatus}.
     */
    private void assertStatusBarContent(String expectedSaveLocation, String expectedSyncStatus) {
        assertEquals(expectedSaveLocation, statusBarFooterHandle.getSaveLocation());
        assertEquals(expectedSyncStatus, statusBarFooterHandle.getSyncStatus());
        guiRobot.pauseForHuman();
    }

}
